% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{thompson2015_se}
\alias{thompson2015_se}
\title{Breast Cancer Cell lines PMC42-ET and PMC42-LA from Cursons et al. (2015)}
\format{
A SummarizedExperiment object, containing gene expression data of
different sub-lines of human breast cancer cell lines. The \code{\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::colData()}}
function can be used to access the sample annotations.
}
\description{
Unpublished dataset from the laboratory of Rik Thompson at the Queensland
University of Technology (QUT).
Dataset used in Cursons et al. 2015. where parallel ‘deep sequencing’ of RNA
(RNA-Seq) were conducted to examine the changes in expression profiles
between PMC42-ET and PMC42-LA cells. Cells were treated for 3 days in the
presence or absence of 10 ng/ml EGF. ONLY the controls are processed. There
are 3 biological replicates per condition which were summed. Sequenced on the
Illumina HiSeq 2000, 100bp paired end.
}
\details{
Data from this publication were downloaded from SRA and/or EMBL-EBI ENA and processed into a
SummarizedExperiment object. Sample annotations were modified from the original publication.
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
emt_datasets <- query(eh, "emtdata")
}
\references{
Cursons, J., Leuchowius, KJ., Waltham, M., Tomaskovic-Crook, E., Foroutan, M.,
Bracken, CP., Redfern,A., Crampin, EJ., Street, I., Davis, MJ. & Thompson,EW. (2015)
Stimulus-dependent differences in signalling regulate epithelial-mesenchymal plasticity and
change the effects of drugs in breast cancer cell lines. Cell Commun Signal 13, 26 (2015).
}
